pfg-agent/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pfg-agent"
version = "0.1.0"
description = "Autonomous AI agent for Prompt for Good — claims GitHub issues and opens PRs"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.11"

dependencies = [
    "anthropic>=0.29.0",
    "langchain>=0.2.0",
    "langchain-anthropic>=0.1.0",
    "gitpython>=3.1.43",
    "PyGithub>=2.3.0",
    "httpx>=0.27.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.3.0",
    "structlog>=24.2.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-mock>=3.14.0",
    "ruff>=0.4.0",
]

[project.scripts]
pfg-agent = "pfg_agent.main:main"

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["E501"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
